decouple log operation logic
